test: align set_schema_test.lua with core exclude_null changes
The fix for tarantool/tarantool#8649 always sets
index_object.parts.exclude_null to false when it is actually false,
however the test set_schema_test.lua fails [1], because it expects
index_object.parts without the exclude_null field. This patch fixes
the test so that it allows both nil and false for exclude_null.

1. https://github.com/tarantool/tarantool/actions/runs/5889232292/job/15972202950